worker/requirements.txt
temporalio>=1.5.0
python-dotenv>=1.0.0
